Water Hider + Breakable Box
I know the Breakable Box was not supposed to be created atm, but it has been, and player bullets have been edited to dissapear when hitting them, taking out a special hitpoint counter for it.

In addition, the Water Hider enemy has been created, but needs sprites for hit-stun, hiding, and idle.  Other than that, hi mechanics are working as intended.